Destroy subprocesses on shutdown
The song and dance with sh/sh-without-orphans is to keep
the public fn `sh` unaltered (though maybe it doesn't need
to be?). It might be nice to be able to scope the killings
to particular eval-in :subprocess calls, e.g. lein repl, but
that would take major surgery. In particular, it'd mean
adding a parallel set of defmethods to match the other
eval-in possibilities.

refs #455